dmd.dmodule: Use associative array for deduplication
Fix an O(n^2) inefficiency in the Dsymbol deduplication logic in the
addDeferredSemantic, addDeferredSemantic2 and addDeferredSemantic3
functions.

The duplication check was added in
dc9fa6c08c9de1610fbdb7b7c1e033fd1518bd89. Unfortunately, one effect of
this is that each Dsymbol insertion required a full linear scan of the
Dsymbols array. In modules which cause a lot of templates to be instantiated,
this can significantly impact compilation times.

Resolve this by switching to an AssocArray as the mechanism for
deduplication.